requirements.txt
django==2.2.10
djangorestframework==3.11.0
requests==2.23.0
idna==2.9
importlib-metadata==1.6.0
isort==5.6.4
